ocp4: Add machine config for ElasticSearch
As Elasticsearch requires sysctl settings to be properly set for
startup, we need to use the OCP 4 concept of a [machine config][mc] to
set those values.  This is a working, minimal configuration that will
correctly create the sysctl dropin file that quadruples the maximum
number of memory maps for kernel allocation.